0x0F-function_pointers

+++ 0x0F-function_pointers/1-array_iterator.c
@@ -1,9 +1,15 @@
 #include "function_pointers.h"
 /**
+ * array_iterator -  executes a function given as
+ * a parameter on each element of an array.
+ * @array: pointer to ints
+ * @size:size of array
+ * @action: pointer to a function to exectute
+ * Return: void
  */
 void array_iterator(int *array, size_t size, void (*action)(int))
 {
-	unsigned int i;
+	size_t i;
 
 	if (array == NULL || action == NULL)
 	{
